example/requirements.txt
easy_thumbnails==2.7.0
WebTest==2.0.35
django-webtest==1.9.7
coverage==5.3
isort==5.5.3
pytest==6.0.2
pytest-cov==2.10.1
pytest-django==3.10.0
pytest-env==0.6.2
pytest-factoryboy==2.0.3
pytest-html==2.1.1
pytest-isort==1.2.0
pytest-black==0.3.11
